// src/client/toteco/models/product.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Menu } from './menu';
import { Publication } from './publication';
 /**
 * 
 *
 * @export
 * @interface Product
 */
export interface Product {

    /**
     * @type {string}
     * @memberof Product
     */
    id?: string;

    /**
     * @type {string}
     * @memberof Product
     */
    name?: string;

    /**
     * @type {number}
     * @memberof Product
     */
    created?: number;

    /**
     * @type {number}
     * @memberof Product
     */
    updated?: number;

    /**
     * @type {boolean}
     * @memberof Product
     */
    inMenu?: boolean;

    /**
     * @type {number}
     * @memberof Product
     */
    price?: number;

    /**
     * @type {number}
     * @memberof Product
     */
    score?: number;

    /**
     * @type {Publication}
     * @memberof Product
     */
    publication?: Publication;

    /**
     * @type {Menu}
     * @memberof Product
     */
    menu?: Menu;
}