// src/client/toteco/models/product-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ProductDTO
 */
export interface ProductDTO {

    /**
     * @type {string}
     * @memberof ProductDTO
     */
    name?: string;

    /**
     * @type {boolean}
     * @memberof ProductDTO
     */
    inMenu?: boolean;

    /**
     * @type {number}
     * @memberof ProductDTO
     */
    price?: number;

    /**
     * @type {number}
     * @memberof ProductDTO
     */
    score?: number;

    /**
     * @type {string}
     * @memberof ProductDTO
     */
    menuId?: string;

    /**
     * @type {string}
     * @memberof ProductDTO
     */
    publicationId?: string;
}